+++ frontend/src/utils/matchEvents/matchHeaderHelper.ts
@@ -1,5 +1,5 @@
 import type { Match } from '@/types/match';
-import type { MatchEvent } from '@/types/matchEvents';
+import type { MatchEvent } from '@/types/matchEvent';
 
 export function formatMatchDate(kickoffTime: string) {
   return new Intl.DateTimeFormat('en-US', {
